+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Show only first line of the last message in ConversationCard [#4545](https://github.com/rokwire/illinois-app/issues/4545).
+- Detect URL links in text body of conversation message card [#4547](https://github.com/rokwire/illinois-app/issues/4547).
 ### Fixed
-- [BUG] Additional bug fixes for Messages [#4548](https://github.com/rokwire/illinois-app/issues/4548).
+- Additional bug fixes for Messages [#4548](https://github.com/rokwire/illinois-app/issues/4548).
+
 ## [6.1.30] - 2024-12-19
 ### Changed
 - Message Conversation UI Updates and Bug Fixes (in progress) [#4536](https://github.com/rokwire/illinois-app/issues/4536)